Rename forest_benchmarking import to forest.benchmarking
