+++ forest/benchmarking/classical_logic/ripple_carry_adder.py
@@ -28,9 +28,9 @@ from pyquil.quil import Pragma
 from pyquil.api import QuantumComputer
 from pyquil.unitary_tools import all_bitstrings
 
-from forest_benchmarking.readout import _readout_group_parameterized_bitstring
-from forest_benchmarking.classical_logic.primitives import *
-from forest_benchmarking.utils import bit_array_to_int, int_to_bit_array
+from forest.benchmarking.readout import _readout_group_parameterized_bitstring
+from forest.benchmarking.classical_logic.primitives import *
+from forest.benchmarking.utils import bit_array_to_int, int_to_bit_array
 
 
 def assign_registers_to_line_or_cycle(start: int, graph: nx.Graph, num_length: int) \
