+++ forest/benchmarking/operator_estimation.py
@@ -9,8 +9,8 @@ from pyquil.paulis import (PauliSum, PauliTerm, commuting_sets, sI,
                            term_with_coeff, is_identity)
 from pyquil.quil import Program
 from pyquil.gates import RX, RY, RZ, MEASURE
-from forest_benchmarking.readout import estimate_confusion_matrix
-from forest_benchmarking.compilation import basic_compile
+from forest.benchmarking.readout import estimate_confusion_matrix
+from forest.benchmarking.compilation import basic_compile
 
 STANDARD_NUMSHOTS = 10000
