+++ forest/benchmarking/plotting/hinton.py
