+++ tests/test_rpe.py
@@ -1,12 +1,13 @@
 import numpy as np
+from forest.benchmarking.compilation import basic_compile
 from numpy import pi
-from pyquil.gates import H, RY, RZ
-from forest_benchmarking import rpe
-from pyquil.noise import damping_after_dephasing
 from pandas import Series
+from pyquil.gates import H, RY, RZ
+from pyquil.noise import damping_after_dephasing
 from pyquil.quil import Program
 from pyquil.quilbase import Measurement
-from forest_benchmarking.compilation import basic_compile
+
+from forest.benchmarking import rpe
 
 
 def test_state_prep(wfn):
