+++ tests/test_t2.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from forest_benchmarking.qubit_spectroscopy import fit_to_exponentially_decaying_sinusoidal_curve
+from forest.benchmarking.qubit_spectroscopy import fit_to_exponentially_decaying_sinusoidal_curve
 
 
 @pytest.fixture()
